SharedPreferences Update
A new interface, SettingsInterface, allows all classes that implement
the interface to access and modify the shared preferences of the game.

Both the GameSettingsActivity and the SheepHerderActivity initialize
default settings for the fox, sheep, and dog entities. The
GameSettingsActivity also allows the user to edit these preferences by
dragging a seekbar.