SharedPreferences Bug Fix
This commit fixes a bug where changes in the GameSettingsPanel did not
propagate to the rest of the class. This is because two different
preference files were used. Both of these preferences files were
combined into one file, which is created upon initialization of the main
activity and can be changed by the user only through the
GameSettingsActivity.

This commit also features threading in the MainActivity. Preferences are
loaded in a thread separate from the UI drawing.