Multiple Clicks Replaced With Long Press
Dog now listens for a long press on the touch screen before moving to a
new position. This required the moveTo() method for the dog to be
overridden to check the specific tapped pixel and possibly make
necessary corrections to it so the dog doesn't get trapped in a corner.